refactor(console): extract getRoleMap to shared utility and scope CSS selectors
Extract duplicated getRoleMap helper to DeleteAccountModal/utils.ts
and import from both the modal and the new page component.

Scope bare p/p+p and .actions CSS selectors under .container to prevent
global style leakage from the CSS module.
